unbind events when clearing model manager
there was a bug where if you called:

    MyResource.models.clear();

it would clear the manager's models, but it would still have events
bound to them. then if you had a collection that contained a model that
was previously in the manager, and that model triggered a change event,
the collection would receive two change events from the manager, one
from before clearing it, and one from after.

as the long description suggests, this would be hard to test for, so i'm
just committing a fix.