correctly serialize nested changes
if you set the 'foo.bar' model of a property, then there's a change
recorded for both 'foo' and 'foo.bar'. when you later go to serialize
the changes for an update request (i.e. calling 'save' on an existing
model) this makes sure that we only serialize 'foo.bar' and not 'foo' in
addition.

pertains to DAQ-604