WIP return deferred from `Model#set()`
this is part of making it possible to validate `Model#set()` calls.
pretty much every case where we'll want to validate those set calls,
we'll want some means of providing user feedback, so 'set' can't just
fail silently, it's got to provide the instance of
`fields.ValidationError` so that (some day) binding libraries can use
that to automatically display error messages.

the best way to provide this error information is through a deferred
since `Model#validate()` returns a deferred as well, and exceptions are
a pain to deal with.

pertains to #24